refactor: capitalize Import folder, replace enum with const object
