--- apps/antalmanac/src/components/Header/import/types.ts
@@ -1,6 +1,0 @@
-export enum ImportSource {
-    ZOT_COURSE_IMPORT = 'zotcourse',
-    STUDY_LIST_IMPORT = 'studylist',
-    AA_USERNAME_IMPORT = 'username',
-    JSON_IMPORT = 'json',
-}
